fix: icon not shown for negative `tag` and `badge`, wring warning icon for Orange theme (#1216) (#1219) (#1226)
Fix missing icon in negative status for tag and badge components.
Fix warning icon for Orange theme.

Closes #1216
Closes #1219

+++ OUDS/Core/ComponentsUIKit/Sources/Indicators/_OUDSBadge.swift
@@ -48,7 +48,7 @@ extension OUDSUIKitBrige {
     ///    - accessibilityLabel: The accessibility label bringing meanings about the count value
     ///    - status: The status of this badge, default set to *neutral*
     ///    - size: The size of this badge, default set to *medium*
-    @MainActor public static func createBadge(count: UInt,
+    @MainActor public static func createBadge(count: UInt8,
                                               accessibilityLabel: String,
                                               status: OUDSBadge.Status = .neutral,
                                               size: OUDSBadge.IllustrationSize = .medium) -> UIViewController
